Updated openstack/openstack Project: openstack/nova-specs  c58a68456a7350eb91e14113cb57bd43d909bdc6
VMware: Native HTML5 consoles for VMware

This is a proposal to add support for the native HTML5 console available in
vSphere 5.5 and later (referred as WebMKS console). This will solve some of the
current problems that we have with VNC consoles with respect to deployment,
scalability and security. The existing support for VNC consoles will remain but
users will be encouraged to use WebMKS when possible.

blueprint vmware-webmks-console

Change-Id: Ie8e4d6f1091e2e48848952d6e0018e20990188aa