Updated openstack/openstack Project: openstack/requirements  9c2443e490ea53034c295e19997e1b8e1cae26e9
Bump futures up to 3.0 or greater

The 3.0 release of futures removed support
for py2.5 and py3.1 and also enables the ability
to correctly retain the traceback of exceptions
on 2.x versions of python. This functionality
will enable much better exception tracking when
the futures/executor interface is used and would be
a welcome addition in futures using projects (and
helps aid the migration to 3.x which includes
tracebacks by default).

This release also fixes memory leaks, as_completed
and other various bugs that were previously occurring
with the 2.x branch.

The new home of this code also moved from google
code to https://github.com/agronholm/pythonfutures

Change-Id: I36f1241b983f6552d8bd0471e6a7485532a95a14