Updated openstack/openstack Project: openstack/python-keystoneclient  3f757656a485d3cec1a56fe4d3f9f07a313a74c7
Use 'id' instead of 'protocol_id' in federation protocol tests

The actual attribute returned in object references of the /protocols API
is 'id', as in all other keystone APIs that return objects. The
implementation of new_ref() here doesn't actually include an 'id'
reference though, and goes out of it's way to test the wrong thing. This
patch fix that, eliminates the workarounds, and does a touch of
refactoring to bring these tests in line with tests of other client
managers.

Change-Id: I9a272b3ef91934e780106d89b5091b4bfb87ad29
Closes-Bug: 1453847