Updated openstack/openstack Project: openstack/ironic  1f8dad5a69102b70c986212120b96740cd4f5049
Check temp dir is usable for ipmitool driver

This patch adds directory check functions to common/utils.py.
It adds a call to the new check functions in ipmitool init.
So we can test the temp directory, upon startup to ensure that
the directory is usable.

These pre-flight checks do not change the need for valid
and accurate error handling, They are to quickly inform
an operator that there is a configuration error that will
prevent the driver from performing as expected.

Also adds a missing unit test for the ipmitool drivers power
validate function "test_power_validate".

Change-Id: Iefc75203ccde1d25fd91c9cb815d871b068255a2
Partial-Bug: #1428722