Updated openstack/openstack Project: openstack/oslo.messaging  e1e2abaed8cf9fcfbc8698005aad6936d050b4f2
rabbit: remove unused consumer interfaces

The consumer code is over engineered, it allows to override
everything, but the override is always done with functools.partial.

None of the child Class have the same signature, sometimes
the constructor use the parameter name as the parent class but for
a different purpose, that makes the code hard to read.

It's was never clear which options is passed to the queue and the
exchange at this end to kombu.

This changes removes all of that stuffs, and only use the kombu
terminology for consumer parameters.

Alse we don't hardcode anymore the tag and the channel in the consumer
class, to allow to change them without recreating a consumer object
in the futur.

Change-Id: Ie341f0c973adbda9a342cb836867345aa42652d1