Updated openstack/openstack Project: openstack/taskflow  b671f2e162cbaf6322c088c66218f189ec5f6420
Just let the future executors handle the max workers

Instead of providing and retaining a thread count in
the worker and action engine executors and checking it
and handling the none case, we can just let the future
types handle this already (which they already do). And
when displaying this information in the worker banner
use a new future executor attribute that is the maximum
number of workers that will be ever created.

Change-Id: I765c22936b53cdbb8a90195a764d4c67bcc3f34b