Updated openstack/openstack Project: openstack/taskflow  82103eb553b19a88a7132dd39a48d769d36a49c9
Avoid re-normalizing paths when following links

There is not any need to re-normalize internal paths
once the initial get() or __get_item__() have normalized
there path arguments so in _get_item() just tell the fetching
function to avoid re-normalizing the paths its provided.

This also works for link following since the target
placed in the fake inodes metadata is already normalized
when it is inserted.

Part of blueprint make-things-speedy

Change-Id: If31d61b744f8df618c7db3afffc4a118cb17003a