Updated openstack/openstack Project: openstack/neutron  39b90888cf842ed145ac81bb01b97d9bee476284
Make it clear the rfe tag is lower-case

This tripped me up, I tried adding RFE to my bug and was denied.
It seems the tag was added with lower-case letters.  This is fine with
me as it is consistent with most, if not all, other tags in the
project.  If it stays lower-case, we should at least make the quoted
tag in the policy document lower-case.

Change-Id: I9c72a2db2a168b56b1137839f5bfc1d2068d9f0c