Updated openstack/openstack Project: openstack/heat  4032e40cce01cf2f31a359d9bbbcb34c3f0d2cae
Don't use scoping information for admin auth

The only place that admin_auth seems to be used here is gaining the
user_id for establishing a trust. Fetching a token for only the user_id
is not an operation that requires scoping to a project so request an
unscoped token here.

Change-Id: Id7c30ec2340addb30d5bd463891c9fdd9280009e