Updated openstack/openstack Project: openstack/requirements  f978c9dc4987f8f954c3bc4cdb9f0c987fc5584a
global-requirements: Add ovs

ovs is the Python library for interacting with Open vSwitch.  Neutron
already uses this (starting in the Kilo release) but it was only just
now added to PyPI.

Info "For new Requirements" from the README:

- Is the library actively maintained?

Yes.  It's included in the main openvswitch git tree and is maintained
by the ovs team.

- Is the library good code?

Yes.

- Is the library python 3 compatible?

More testing is required as it's not clear what testing has already
been done.  It has no additional dependencies and I'm not aware of
anything else that would make Python 3 compatibility difficult.

- Is the library license compatible?

Yes.  Apache 2

- Is the library already packaged in the distros we target (Ubuntu
  latest / Fedora latest)?

Yes.  This is already an OpenStack dependency.  It was just only
recently uploaded to PyPI to make testing in our CI easier.

- Is the function of this library already covered by other libraries
  in ``global-requirements.txt``?

No.

Change-Id: Ieb40dce8448d8d10150540238e2ab8dded5bc743