Updated openstack/openstack Project: openstack/keystonemiddleware  de7a54efd325764dd79a0ca07d77ca42281d2bcf
Add keystone v3 API to fetch revocation list

Currently in auth_token middleware, only the v2 keystoneclient API is
used to fetch the revocation list. However, the v3 keystoneclient API
should be implemented and used as well. There should be a check on the
value of auth_version, and the corresponding API should be called.

Change-Id: I1c48da9afadfbdae2f69820e0b89bc02ec54000e
Closes-Bug: #1451445