Updated openstack/openstack Project: openstack/nova  3ac6b922cdbe46fd81d2791d00bb452aec33e1a1
Move our ObjectListBase to subclass from the Oslo one

This lets us drop a lot of the stuff in our object list base class
in favor of what is provided by oslo. Note that a couple of methods
have to be added in there temporarily because they are on the oslo
base object class, which we don't yet inherit from.

Related to blueprint use-oslo-objects

Change-Id: If5514804f6cd2f54eb103e887575f036edebfcee