Updated openstack/openstack Project: openstack-infra/puppet-cgit  0ae79e0106746475d72cedef72486d677f70f708
New Updated config for CentOS 7 and Apache 2.4

This patch takes the original which was reverted and
adds mod_version which is needed to parse IfVersion if
on a Debian system.

When using puppet-cgit under CentOS 7, apache
fails because the config file is not working
in Apache 2.4 version.
Also, as CentOS 7 is starting apache on systemd,
it needs to load the systemd apache config under
conf.modules.d.

Original Patch: I7e0d51ee176c4f27721c16afeaae120eb8edf7ab

Change-Id: If3acc672ccd85b5704a2120379b60cb95528b7f7
Co-Authored-By: Yolanda Robla <yolanda.robla-mota@hp.com>
Co-Authored-By: Nicola Heald <nicola.heald@hp.com>