Updated openstack/openstack Project: openstack/neutron-vpnaas  23365d39399e046b0358b06769116e44f5d8da1f
Remove contextlib.nested from tests

With the deprecation of this method in Python2.7, the neutron hacking
check has been modified to check for this (N324). Since this check has
been upstreamed (I171fbdb89892a3d4548bf2ca52f4a7dd9ef8dccb), and the
neutron-vpnaas repo uses the same hacking check, the VPN test cases
are now failing.

This commit does the same changes as was done in neutron to remove use
of this method.

Change-Id: Ie490e89c1a65e126ad0a57d062076e472762837e
Partial-Bug: 1428424