Updated openstack/openstack Project: openstack/nova  195b1a667bb198fa3c9f35b12467c0935b6a322d
Drop compute RPC 3.x support

This drops all the things we no longer need when not supporting the
compute RPC 3.x interface. The major changes are removing of the
run_instance() path and the mass of tests that hit that route. There
are also quite a few tests that still depend on object_compat
decorations on compute manager methods, which require some
changes here.

Change-Id: I60a89bf7884dbcb066c63101168b359013fe832b