Updated openstack/openstack Project: openstack/nova  054cbb7baf1abadb400722352a03cfa841a20448
libvirt: safe_decode xml for i18n logging

The xml argument passed to _create_domain can be a utf-8 encoded string
which causes a UnicodeDecodeError when it is substituted into the _LE
unicode translated message. Safely decoding the xml argument before
attempting to substitute it into the error message avoids the
UnicodeDecodeError.

Closes-Bug: #1453274
Change-Id: I4cf1836f4ca9097f7c6d98c5212a14d24111fe67