Updated openstack/openstack Project: openstack/oslo.rootwrap  e3d9a86a36d7c93a266ce83f250352a4d3ec6b06
Generate a oslo-rootwrap console script

Currently projects like nova, cinder, neutron etc are generating
their own project-rootwrap script, but they are just copies of each
other as we send the configuration file as a parameter. So we should
generate a script in oslo.rootwrap itself and get rid of others
over a period of time. In this change, we also add a script for the
daemon mode.

Change-Id: I9e9033fbf61161149e586098e4a60b9522ec5437