Updated openstack/openstack Project: openstack/oslo.config  8aa2947ef6245d355159396f125f30897f5c51f9
Allow loading custom config files into Config Fixture.

This patch adds a set_config_files option that allows a tester to
force load a specific list of configuration files. It will override the
existing set of files, and trigger a reload of those configuration
files.

Note: This _does_ trigger the autodetection mechanism, as the
present state of oslo_config makes it so __call__ is the only thing
that sets self._args. After triggering, however, it immediately
overrides the detected file list and replaces it with the provided one.

Tests provided.

Change-Id: I0bcea8be9d2527c61bd8d150e9ed606fabdd290e