Updated openstack/openstack Project: openstack/oslosphinx  6a1260bcc8abd81f7af2dd66f0f70dbc3b77b759
Remove the unneeded horizontal scrollbar

Even on really large screens, the horizontal scrollbar is shown. This is
because the `div.related` class has a width of 100% in basic.css and in
tweaks.css a border is also defined. This causes the div to be two pixes
wider than the screen.

Change-Id: I2881d8eca99988ff8372b22aecc1a1a90c66a314