Updated openstack/openstack Project: openstack/nova  109b681e9fa81875a6e4942dc2a9b244e8c713e9
Move ebrctl to compute.filter.

Currently ebrctl is network.filters, but the utility is
used by libvirt in nova compute, as part of a Mellanox vif driver.

The ubuntu packages install network.filters with the nova-network package.
But this vif driver only works with Neutron.

To help fix that, this patch move it to compute.filters to ensure it is
always installed with Nova.

Closes-Bug: #1453418

Change-Id: Ib61a565c10edfa662c15fac339b2cbd9bb673631