Updated openstack/openstack Project: openstack/tempest  1c6d87c95093a8d3a9d2d197b3887dd727aa9ef7
Add new config option to globally trigger resource validation

This commit adds a new config option to the config group to trigger
all the validation code. It also deprecates the run_ssh option in the
compute group since this will eventually be used globally. The intent
of this option is so that it will trigger both the creation of any
validation resources as well as used as run_ssh was originally
intended, to trigger ssh into nodes when needed.

Closes-Bug: #1453936
Partially-Implements: bp ssh-auth-strategy
Change-Id: I7d04f81457fc564ed82872e602746c8162fef24c