Updated openstack/openstack Project: openstack-infra/project-config  77cc532f7e9e149d7f68459bf1b3b121326a89de
Make neutron-lbaas-dsvm-api non-voting.

This job is broken.  A patch merged into neutron which breaks this job
because this job isn't actually using the neutron patch it's testing
against:

    https://review.openstack.org/#q,8db41f04d54526104920f3a160203ecf7ef453b0,n,z

The revert of that patch won't merge because the lbaas-dsvm-api job is
installing neutron master.

    https://review.openstack.org/#/c/182413/1

This job should be non-voting until it's fixed to properly test the
right neutron patch(es) instead of neutron master.

Change-Id: Ieca799e069696b460aef8caebda84428643ee259