Updated openstack/openstack Project: openstack-dev/hacking  0e7f484d9fa107831d5f6d8c0daf6760da372dc3
tests: allow to run the same test for different code

If a function is registered to report different tests code (i.e. it
yields both H666 and H667), the test generator will fail as it does not
include the code number in the function name – therefore the function
will be generated twice with the same name. This patch fixes that.

Change-Id: I5394a066250f97882e895873164355e405b9c7cf