Updated openstack/openstack Project: openstack-infra/pypi-mirror  989780da2dfe26d63eac49a09cec5164febbad14
If we specify a branch, always use it

There's a check to make sure that we only run
mirroring against interesting branches by filtering
out any that aren't HEAD or don't start with
remotes/origin, but if we specify a branch on
the command line, we want that branch no matter what.
This is primarily so we can run mirroring tests
on proposed changes to openstack/requirements.

Change-Id: I19cd2b032b6219aed972abb95a180e3d402d529e