Updated openstack/openstack Project: openstack/neutron  bc28aa7f361bd3c7243e9e475e98099e93ec0cf0
Update build hooks.

The pbr setup_hook has not been needed for a while, so remove it.  The
neutron hook has been broken for a while: it places a setup_requires
build dependency on everything in neutron/__init__.py, which is
non-empty, but setup_requires is handled by easy install so we try
very hard to avoid it. Instead, we can use environment markers to
selectively include the win32 dependencies without requiring that
neutron be importable during setup.py execution. This is unusual
in OpenStack and will eventually be moved to a regular
requirements.txt dependency with the same marker - once we've finished
the integration work to make markers work properly in
requirements.txt.

Change-Id: Icdc403a3ccf06daeccf2a907a7bfeafd8dbbb5dd