Updated openstack/openstack Project: openstack/cinder  8c0f5d048c7142df1f8c3d087a63606523219054
Removing sleep between when a command is sent and 'YES' is sent

Removing sleep between when a command is sent and 'YES' is sent to
confirm the operation. We believe it is added because the Cisco CLI
is mirrored based on Brocade. Removing would increase performance when
launching concurrent vms.

Closes-Bug: #1447776

Change-Id: I3102922456fa9dde3a63ec623565a3427e9b357d