Updated openstack/openstack Project: openstack/tempest  66b9ec5bc978f018d326f2f91188908e4983079a
Fix raise syntax in test.py for python3 compat

The raise syntax was changed in python3 so the usage of passing in
the separate components as different params to raise will not work
when running on python 3. This commit updates it to use a syntax
which is compatible with python3.

Change-Id: I5cb9979e8ba8e324efb95feec33c23238dffade5