Updated openstack/openstack Project: openstack/neutron  c069ada35e69c1c917ea93688b0afca80b4a1794
Consider all address scopes in init_l3

Currently init_l3 retrieves the list of global addresses from the kernel
on a specific device in a network namespace. If any of the addresses are
not in the ip_cidrs argument to init_l3, they will be deleted. The
problem with only listing global addresses is that if a site-local or
link-local address is added during a subnet-create, and the user wishes
to later delete the address, init_l3 will never consider that address
for deletion.

To fix this, init_l3 should not limit its scope when listing addresses
on an interface. It should, however, ignore the default IPv6 link-local
address assigned by the operating system as this address is not known to
Neutron and should not be deleted.

Change-Id: I3d7a3e318e32acae3836c51e4e2e95ae756e645b
Closes-Bug: #1437499