Updated openstack/openstack Project: openstack/oslo.messaging  4fa7795f42a1efcc39c023229589d54ae8670280
rabbit: smarter declaration of the notif. queue

The NotifyPublisher was redeclaring again and again,
the same exchange and queue each times a notification is sent.

This change fixes that by caching the already declared exchange
and queue for each channel.

Also, to make the test pass. 'Connection.ensure' have been updated
to have the same behavior for amqp and memory driver about
kombu recoverable_errors. And the hostname and port of the memory
driver are set to not fail when we print a log message.

Closes bug: #1437902

Change-Id: I20d133ac67b8a8a4c51d51b6a1b2369aa44ffe2f