Updated openstack/openstack Project: openstack/diskimage-builder  df437f9b7cecf02a16b4515dae495143fc79c9b0
Add packages required for iscsi extension in agent

In Kilo, we added added an iscsi extension in
ironic-python-agent which requires tgtd and tgtadm.
This commit adds scsi-target-utils to the ramdisk for
this.

Also the git protocol for retrieving the source code is
changed to http.  Git protocol doesn't go through a proxy,
but http can.

Closes-Bug: 1449854
Change-Id: I8cf274913a16404941770d0c6115bd6feec1ccb8