Updated openstack/openstack Project: openstack/cinder  cf7da9cb4f6b4e7c922960c3f561f255c5e07b47
Fix incorrect reraising of exceptions

There are several places in the code where exception handling raises
the exception variable rather than just calling raise. This results
in the traceback being incorrect.

Had considered adding a hacking check for this, but that becomes a
little tricky. There are valid places where "raise ex" is used that
would prevent a simple check.

Change-Id: Ib2bd745c7ef600c514a94c3fd638d15d17a623a2