Updated openstack/openstack Project: openstack/cinder  80a71224e0fce3d2ad6e90262615d4be4721018c
Switch to oslo_versionedobjects

oslo_versionedobjects was not made available until very
late in the Kilo cycle (i.e. near the end of kilo-3).
In order to make progress on cinder objects, a fork of
nova objects was made, so that proper trial and testing
could be done. The following patch makes the switch to
use oslo_versionedobjects.

Implements: blueprint cinder-objects
Change-Id: I883f387c8247e8d79da82016a624cef2180cde88