Updated openstack/openstack Project: openstack/cinder  d99b9196838fcf411b6572bdab6a2cdbcf867090
Cinder os-force_detach api returns 500

If invalid 'attachment_id' is passed to the os-force_detach
api, then it returns 500 error because even though
VolumeAttachmentNotFound exception is raised from manager,
it gets converted to RemoteError due to rpcapi call.

Caught RemoteError exception in the api and returned 400
error to the user if exception type is VolumeAttachmentNotFound
or InvalidVolume. In other cases 500 error is raised.

Closes-Bug: 1450431
Change-Id: I5e308d9f4b3772284caaedb3d2d0b3f7d25e1e5d