Updated openstack/openstack Project: openstack-infra/tripleo-ci  10304a574d2baec3aa3e711f2c32a947ddf08053
Add a ceph job

This patch will allow us to create a job that uses Ceph
RBD for Nova, Cinder, and Glance.

For now we are making use of the fact that Ceph is
installed in the compute image. Eventually we'll
want to make use of a more flexible way to build
and load images within devtest (using either a single
image for everything, or building a new Ceph specific
image)

Change-Id: I22c9037fb734970396dd5045b8adf8a979bcf614
Depends-On: Id6a3fbc3477e85e8e2446e3dc13d424f9535d0ff
Depends-On: Ic47d028591edbaab83a52d7f38283d7805b63042