Updated openstack/openstack Project: openstack/sahara  f81454dce998e78bcb77cd75ddd438bac9ca17ee
[HDP] java64_home not pointing at default-installed JDK for plugin

At this time HDP's java64_home property defaults to /opt/jdk1.6.0_31.
However, the now-default openjdk element installs to /usr/lib/jvm/java-openjdk.
If nodes are connected to the internet at cluster spinup time, this does not
cause an overt error, but querying for JDK location will return error code 2,
resulting in no -j argument being passed to 'ambari-server setup -s', which in
turn will result in a download of Oracle's JDK and successful installation. In
the case that the nodes cannot connect to the internet (which should be a
success case with a pre-installed VM image,) cluster creation fails at this
stage.

This patch updates java64_home to the default JAVA_HOME from the openjdk
element.

Change-Id: I1d53db67341626190e7240f0ae4165f91d364a2c
Closes-bug: 1453165