Updated openstack/openstack Project: openstack/ironic-python-agent  8a94a8979c0717ec345487e1762c5e9089010d4a
Use mkdtemp for mounting virtual media device

When using virtual media, restarting ironic-python-agent
leads to errors like 'File exists'.  This is because
we use static directory (/vmedia_mnt) as mount path
and we don't remove it. It's better to use mkdtemp to create
a temporary directory for mounting virtual media
device.

Change-Id: Ibb4c3d3a2024cf51ef198bd396fb9e86ad72a320