Updated openstack/openstack Project: openstack-infra/subunit2sql  a1ba4871892a78a733449054f44bd21de78eedf7
Add db api method to get uuid from test_id

This commit adds 2 new db api methods to deal with taking the test_id
column and finding the value of the matching id. The first method will
return the id value for a given test_id, while the second returns the
list of all ids in the tests table.

Change-Id: I696aebd499ed47a20af0c13415fc568f2b54f188