Updated openstack/openstack Project: openstack/os-net-config  90f15e38f1e4d80b81f4fa1444dde717917ef8ad
ifcfg: don't set VLAN=yes if is an ovs_port

When using the ifcfg provider vlans on bridges (or bonds)
fail to startup correctly. This is because VLANs on an OVS
bridge are configured as internal ports and these should not
have the VLAN=yes or PHYSDEV ifcfg settings specified.

Change-Id: Ieb743eda8f0a25cdaca43bdea767af26bf504f15
Closes-bug: #1453904