Updated openstack/openstack Project: openstack/heat  3adbe6ad2b5ff9df2bc706647f12eba036273591
Don't raise error if actions are mismatched

Heat stack_resource has similar check for actions, but currently it does
not raise error if actions are different. OS::Heat::Stack resource is
similar on stack_resource so it has same issue, when resources require
some time to change state/action and it should not be so strict.

Change-Id: I20c2d3f8900b7393b1f551384805a35d5ef6497d
Closes-Bug: #1454589