Updated openstack/openstack Project: openstack/oslo.i18n  3e3ce417ede41f8ca838c5e202e7527f20bcfbe4
Misplaced parenthesis causing confusion

The current example in the guidelines for using translation markers
has the closing parenthesis including the string as well as the
formatting argument. This has caused some confusion in reviews where
someone incorrectly follows this pattern and ends up with pep8 errors.

This patch corrects the closing parenthesis placement for both examples
in the "Avoid Forcing the Tranlation of Translatable Variable" section.

Change-Id: I5cd80c3ebec57a7e3686cac946427f7b6e6f3d6d