Updated openstack/openstack Project: openstack/nova  cf3733ac0e607cded074468198486fbdf4c760db
Retry a cell delete if host constraint fails

The instance.destroy method has a constraint that fails during a local
delete if the host is not None, because it should be in that case.
Cells calls local_delete when there's no cell_name on the instance, but
when this fails it's likely that we have a host and cell_name now so the
delete method should be attempted again.

Closes-Bug: 1454839
Change-Id: I412fab6f43cf2c64239a0292946c6c3bb9837110